requirements.txt
discord.py[speed]
git+https://github.com/Rapptz/discord-ext-menus
sympy
tabulate
lxml
Pillow
wolframalpha
matplotlib
pytz
